backend/requirements.txt
# Core
fastapi
uvicorn[standard]
python-dotenv

# CrewAI (use latest versions with anthropic support)
crewai[tools,anthropic]

# HTTP Client
httpx

# Testing
pytest
